Support empty lines on saved filed
This patch adds support to ignore empty lines on saved files when
replaying them.

This is not something that will happen with a saved file as it is, but
it may happen if someone copy pastes from a saved file into another and
accidentally leaves an empty line at the end of the file.